% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFelids.R
\docType{data}
\name{DataFelids}
\alias{DataFelids}
\title{Example dataset}
\format{
A list containing: \describe{\item{\strong{$treefel}}{Phylogenetic tree of felids}.
\item{\strong{$landfel}}{list of 2D landmark configuration for each species of the tree}.
 \item{\strong{$PCscoresfel}}{A data.frame containing 83 shape variables for felids' mandible}.
 \item{\strong{$statefel}}{sabertooth - not sabertooth ecomorph categorization for each species}.
  }
}
\usage{
data(DataFelids)
}
\description{
Geometric morphometrics shape data regarding felids' mandible
  and phylogentic tree (\cite{Piras et al., 2018}).
}
\references{
Piras, P., Silvestro, D., Carotenuto, F., Castiglione, S.,
  Kotsakis, A., Maiorino, L., Melchionna, M.,Mondanaro, A., Sansalone, G.,
  Serio, C., Vero, V. A., & Raia, P. (2018). Evolution of the sabertooth
  mandible: A deadly ecomorphological specialization. \emph{Palaeogeography,
  Palaeoclimatology, Palaeoecology}, 496, 166-174.
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro
  Mondanaro, Marina Melchionna, Mirko Di Febbraro, Antonio Profico, Francesco
  Carotenuto
}
\keyword{RRphylo}
