% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOrnithodirans.R
\docType{data}
\name{DataOrnithodirans}
\alias{DataOrnithodirans}
\title{Example dataset}
\format{
A list containing: \describe{ \item{\strong{treedino}}{Ornithodirans
  phylogenetic tree}. \item{\strong{massdino}}{numeric vector of
  ornithodirans body masses}. \item{\strong{statedino}}{vector of
  ornithodirans locomotory type}. }
}
\usage{
data(DataOrnithodirans)
}
\description{
Ornithodirans' body mass, phylogenetic tree and locomotory type
  (\cite{Castiglione et al 2018}).
}
\references{
Castiglione, S., Tesone, G., Piccolo, M., Melchionna, M.,
  Mondanaro, A., Serio, C., Di Febbraro, M., & Raia, P.(2018). A new method
  for testing evolutionary rate variation and shifts in phenotypic evolution.
  \emph{Methods in Ecology and Evolution}, 9:
  974-983.doi:10.1111/2041-210X.12954
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro
  Mondanaro, Marina Melchionna, Mirko Di Febbraro, Antonio Profico, Francesco
  Carotenuto
}
\keyword{RRphylo}
