% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSimians.R
\docType{data}
\name{DataSimians}
\alias{DataSimians}
\title{Example dataset}
\usage{
data(DataSimians)
}
\description{
The output of Procrustes superimposition as performed by the function \code{\link[Morpho]{procSym}} on 9 simians faces and the phylogenetic tree for such species.
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro
  Mondanaro, Marina Melchionna, Mirko Di Febbraro, Antonio Profico, Francesco
  Carotenuto
}
\keyword{RRphylo}
