% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distNodes.R
\name{distNodes}
\alias{distNodes}
\title{Finding distance between nodes}
\usage{
distNodes(tree,n)
}
\arguments{
\item{tree}{a phylogenetic tree. The tree needs not to be ultrametric and fully dichotomous.}

\item{n}{either a single node or a pair of nodes.}
}
\value{
The function returns a data frame with distances between the focal node and the other nodes on the tree (or for the selected pair only).
}
\description{
The function returns the distance between pairs of nodes. The distance is meant as both patristic distance and the number of nodes intervening between the pair.
}
\examples{
data("DataApes")
DataApes$Tstage->Tstage

distNodes(tree=Tstage,n=64)
distNodes(tree=Tstage,n=c(64,48))
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro Mondanaro, Marina Melchionna, Mirko Di Febbraro, Antonio Profico, Francesco Carotenuto
}
