% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overftiRR.R
\name{overfitRR}
\alias{overfitRR}
\title{Testing RRphylo methods overfit}
\usage{
overfitRR(RR,y,s=0.25,trend=FALSE,shift.node=NULL,shift.state=NULL,
trend.node=NULL,aces=NULL,x1=NULL,aces.x1=NULL,cov=NULL,rootV=NULL,nsim=100,clus=.5)
}
\arguments{
\item{RR}{an object produced by \code{\link{RRphylo}}.}

\item{y}{a named vector of phenotypes.}

\item{s}{the percentage of tips to be cut off. It is set at 25\% by default.}

\item{trend}{if \code{TRUE}, the function performs \code{search.trend} on the subsampled tree, searching for trends in either phenoypes or rates through time. If the argument \code{trend.node} is specified (see below), the \code{trend} argument is automatically set to \code{TRUE}.}

\item{shift.node}{the nodes to be tested in \code{search.shift} under the "clade" condition. Notice that the conditions on \code{search.shift} are exclusive, they cannot be tested together.}

\item{shift.state}{the named vector of states to be tested in \code{search.shift} under the "sparse" condition. Notice that the conditions on \code{search.shift} are exclusive, they cannot be tested together.}

\item{trend.node}{the nodes to be tested by \code{search.trend} for trends in phenotypes or rates through time.}

\item{aces}{if used to produce the \code{RR} object, the vector of those ancestral character values at nodes known in advance must be specified. Names correspond to the nodes in the tree.}

\item{x1}{the additional predictor to be specified if the RR object has been created using an additional predictor (i.e. multiple version of \code{RRphylo}). \code{'x1'} vector must be as long as the number of nodes plus the number of tips of the tree, which can be obtained by running \code{RRphylo} on the predictor as well, and taking the vector of ancestral states and tip values to form the \code{x1}.}

\item{aces.x1}{a named vector of ancestral character values at nodes for \code{x1}. It must be indicated if the RR object has been created using both \code{aces} and \code{x1}. Names correspond to the nodes in the tree.}

\item{cov}{if used to produce the \code{RR} object, the covariate must be specified. As in \code{RRphylo}, the covariate vector must be as long as the number of nodes plus the number of tips of the tree, which can be obtained by running \code{RRphylo} on the covariate as well, and taking the vector of ancestral states and tip values to form the covariate.}

\item{rootV}{if used to produce the \code{RR} object, the phenotypic value at the tree root must be specified.}

\item{nsim}{number of simulations to be performed. It is set at 100 by default.}

\item{clus}{the proportion of clusters to be used in parallel computing.}
}
\value{
The function returns a 'list' containing:

\strong{$rootCI} the 95\% confidence interval around the root value.

\strong{$ace.regressions} the results of linear regression between ancestral state estimates before and after the subsampling.

\strong{$shift.results} if \code{shift.state} is specified, a list including for each state the p-values at each simulation ($states) and the percentage of simulations producing significant p-value ($p.states). If the argument \code{shift.node} is specified, the list contains for each node the p-values at each simulation ($single clades) and the percentage of simulations producing significant p-value ($p.single.clades), and the same figures by considering all the specified nodes as evolving under a single rate ($all.caldes.togheter and $p.all.clades.together).

\strong{$trend.results} if \code{trend = TRUE}, a list including slopes and p-values of phenotypes versus age and absolute rates versus age regressions for the entire tree at each simulation ($tree.phenotype and $tree.rates) and the percentage of simulations showing significant p-values for both variables separated by slope sign ($tree.p). If \code{trend.node} is specified, the list also includes the same results at nodes ($node.phenotype, $node.rates and $node.p) and the results for comparison between nodes ($comparison.phenotype, $comparison.rates and $comparison.p).
}
\description{
Testing the robustness of \code{\link{search.trend}} (\cite{Castiglione et al. 2019}) and \code{\link{search.shift}} (\cite{Castiglione et al. 2018}) to sampling effects and phylogenetic uncertainty.
}
\details{
Methods using a large number of parameters risk being overfit. This usually translates in poor fitting with data and trees other than the those originally used. With \code{RRphylo} methods this risk is usually very low. However, the user can assess how robust the results got by applying \code{search.shift} or \code{search.trend} are by running \code{overfitRR}. With the latter, the original tree and data are subsampled by specifying a \code{s} parameter, that is the proportion of tips to be removed from the tree. Internally, \code{overfitRR} further shuffles the tree by using the function \code{\link{swapONE}}. Thereby, both the potential for overfit and phylogenetic uncertainty are accounted for straight away.
}
\examples{
data("DataOrnithodirans")
DataOrnithodirans$treedino->treedino
DataOrnithodirans$massdino->massdino
DataOrnithodirans$statedino->statedino
# Extract Pterosaurs tree and data
  library(ape)
  extract.clade(treedino,748)->treeptero
  massdino[match(treeptero$tip.label,names(massdino))]->massptero
  massptero[match(treeptero$tip.label,names(massptero))]->massptero

\donttest{
RRphylo(tree=treedino,y=massdino)->dinoRates
RRphylo(tree=treeptero,y=log(massptero))->RRptero

# Case 1 search.shift under "clade" condition
    search.shift(RR=dinoRates, status.type= "clade",foldername=tempdir())->SSnode
    overfitRR(RR=dinoRates,y=massdino,shift.node=rownames(SSnode$single.clades))->overfit.SSnode

# Case 2 search.shift under "sparse" condition
    search.shift(dinoRates, status.type= "sparse", state=statedino,
    foldername=tempdir())->SSstate
    overfitRR(RR=dinoRates,y=massdino,shift.state=statedino)->overfit.SSstate

# Case 3 search.trend on the entire tree
    search.trend(RR=RRptero, y=log(massptero),foldername=tempdir())->STtree
    overfitRR(RR=RRptero,y=log(massdino),trend=TRUE)->overfit.STtree

# Case 4 search.trend at specified nodes
    search.trend(RR=RRptero, y=log(massptero),node=143,foldername=tempdir())->STnode
    overfitRR(RR=RRptero,y=log(massdino),trend.node=143)->overfit.STnode

# Case 5 overfitRR on multiple RRphylo
  data("DataCetaceans")
  DataCetaceans$treecet->treecet
  DataCetaceans$masscet->masscet
  DataCetaceans$brainmasscet->brainmasscet
  DataCetaceans$aceMyst->aceMyst

  ape::drop.tip(treecet,treecet$tip.label[-match(names(brainmasscet),treecet$tip.label)])->treecet.multi
  masscet[match(treecet.multi$tip.label,names(masscet))]->masscet.multi

  RRphylo(treecet.multi,masscet.multi)->RRmass.multi
  RRmass.multi$aces[,1]->acemass.multi
  c(acemass.multi,masscet.multi)->x1.mass

  RRphylo(treecet.multi,y=brainmasscet,x1=x1.mass)->RRmulti
  search.trend(RR=RRmulti, y=brainmasscet,x1=x1.mass,clus=0.5,foldername=tempdir())->STcet
  overfitRR(RR=RRmulti,y=brainmasscet,trend=TRUE,x1=x1.mass)->overfit.STcet

}
}
\references{
Castiglione, S., Tesone, G., Piccolo, M., Melchionna, M., Mondanaro, A., Serio, C., Di Febbraro, M., & Raia, P. (2018). A new method for testing evolutionary rate variation and shifts in phenotypic evolution. \emph{Methods in Ecology and Evolution}, 9: 974-983.doi:10.1111/2041-210X.12954

Castiglione, S., Serio, C., Mondanaro, A., Di Febbraro, M., Profico, A., Girardi, G., & Raia, P. (2019) Simultaneous detection of macroevolutionary patterns in phenotypic means and rate of change with and within phylogenetic trees including extinct species. \emph{PLoS ONE}, 14: e0210101. https://doi.org/10.1371/journal.pone.0210101
}
\author{
Silvia Castiglione, Carmela Serio, Pasquale Raia
}
