% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRphylo-defunct.R, R/swap.phylo.R
\name{RRphylo-defunct}
\alias{RRphylo-defunct}
\alias{swap.phylo}
\title{Defunct functions in Package RRphylo}
\usage{
swap.phylo()
}
\description{
These functions are no longer available.
\itemize{
 \item \code{swap.phylo}: This function is defunct. Please check \code{\link{overfitRR}}, instead.
}
}
