% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRphylo-deprecated.R, R/swap.phylo.R
\name{RRphylo-deprecated}
\alias{RRphylo-deprecated}
\alias{swap.phylo}
\title{Deprecated functions in Package RRphylo}
\usage{
swap.phylo()
}
\description{
These functions still work but will be removed (defunct) in the next version.
\itemize{
 \item \code{swap.phylo}: This function is deprecated, and will
 be removed in the next version of this package. Please check \code{\link{overfitRR}}, instead.
}
}
\section{\code{swap.phylo}}{

For \code{swap.phylo}, use \code{\link{overfitRR}}.
}

