% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRphylo-package.R
\name{RRphylo-package}
\alias{RRphylo-package}
\title{Phylogenetic Ridge Regression Methods for Comparative Studies}
\description{
\pkg{RRphylo} provides tools for phylogenetic comparative
  analysis. The main functions allow estimation of phenotypic evolutionary
  rates, identification of shifts in rate of evolution, quantification of
  direction and size of evolutionary change of multivariate traits, and
  computation of species ontogenetic vectors. Additionally, there are
  functions for simulating phenotypic data, manipulating phylogenetic trees,
  and retrieving information from phylogenies. Finally, there are functions
  to plot and test rate shifts at particular nodes.

The complete list of functions can be displayed with
  \code{library(help = RRphylo)}
}
\references{
Castiglione, S., Tesone, G., Piccolo, M., Melchionna, M.,
Mondanaro, A., Serio, C., Di Febbraro, M., & Raia, P.(2018). A new method for
testing evolutionary rate variation and shifts in phenotypic evolution.
\emph{Methods in Ecology and Evolution}, 9:
974-983.doi:10.1111/2041-210X.12954

Piras, P., Silvestro, D., Carotenuto, F., Castiglione, S.,
Kotsakis, A., Maiorino, L., Melchionna, M., Mondanaro, A., Sansalone, G.,
Serio, C., Vero, V.A., & Raia, P. (2018). Evolution of the sabertooth
mandible: A deadly ecomorphological specialization. \emph{Palaeogeography,
Palaeoclimatology, Palaeoecology}, 496: 166-174.
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro
  Mondanaro, Marina Melchionna, Mirko Di Febbraro, Antonio Profico, Francesco
  Carotenuto
}
