% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTree.R
\name{scaleTree}
\alias{scaleTree}
\title{Phylogenetic tree calibration}
\usage{
scaleTree(tree, tip.ages, node.ages=NULL, min.branch=0.1)
}
\arguments{
\item{tree}{a phylogenetic tree. The tree needs not to be ultrametric and fully dichotomous.}

\item{tip.ages}{a named vector of tip ages in terms of distance from the youngest tip.}

\item{node.ages}{a named vector of node ages in terms of distance from the youngest tip. If no calibration date for nodes is available except for the tree root, the vector must contain as only entry the tree height with the root number as name.}

\item{min.branch}{the minimum branch length that will be imposed when no age is specified.}
}
\value{
Rescaled phylogentic tree.
}
\description{
The function is a wrapper around the functions "scalePhylo", "assign.ages", and "assign.brlen" written by Gene Hunt (http://paleobiology.si.edu/staff/individuals/hunt.cfm) . It rescales tree branch lengths according to given calibration dates.
}
\examples{
\dontrun{
library(ape)
library(phytools)
set.seed(76)
rtree(100)->tree

max(nodeHeights(tree))->H

#### Example 1 ####
H-diag(vcv(tree))->tipAges
tipAges[tips(tree,160)]<-min(tipAges)
H->nodeAges
names(nodeAges)<-Ntip(tree)+1

scaleTree(tree,tipAges,nodeAges)->treeS1

edge.col<-rep("black",nrow(tree$edge))
edge.col[which(treeS1$edge[,2]\%in\%getDescendants(treeS1,160))]<-"red"

layout(2:1)
plot(tree,edge.color = edge.col,show.tip.label=F)
plot(treeS1,edge.color = edge.col,show.tip.label=F)

#### Example 2 ####
nodeAges<-c(H,4.5,6,6)
names(nodeAges)<-c(101,143,171,121)
tipAges<-H-diag(vcv(tree))
scaleTree(tree,tipAges,nodeAges)->treeS2

edge.col<-rep("black",nrow(tree$edge))
edge.col[which(treeS1$edge[,2]\%in\%c(getDescendants(treeS1,143),
                                    getDescendants(treeS1,171),
                                    getDescendants(treeS1,121)))]<-"red"

layout(2:1)
plot(tree,edge.color = edge.col,show.tip.label=F)
plot(treeS2,edge.color = edge.col,show.tip.label=F)
}
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro Mondanaro, Marina Melchionna, Mirko Di Febbraro, Antonio Profico, Francesco Carotenuto
}
