% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCetaceans.R
\docType{data}
\name{DataCetaceans}
\alias{DataCetaceans}
\title{Example dataset}
\format{A list containing:
\describe{
\item{\strong{treecet}}{Cetaceans phylogenetic tree}.
\item{\strong{masscet}}{numeric vector of cetaceans body masses (ln kg)}.
\item{\strong{aceMyst}}{body mass (ln kg) for \emph{Mystacodon selenensis}, used as node prior at the ancestor of the Mysticeti}.
}}
\usage{
data(DataCetaceans)
}
\description{
Cetaceans' body mass and phylogenetic tree.
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro Mondanaro, Marina Melchionna, Mirko Di Febbraro, Antonio Profico, Francesco Carotenuto
}
\keyword{RRphylo}
