% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.trend.R
\name{search.trend}
\alias{search.trend}
\title{Searching for evolutionary trends in phenotypes and rates}
\usage{
search.trend(RR,y,nsim=100,clus=.5,node=NULL,cov=NULL,foldername,ConfInt=c(FALSE,TRUE))
}
\arguments{
\item{RR}{an object produced by \code{\link{RRphylo}}.}

\item{y}{the vector (or matrix if multivariate) of phenotypes.}

\item{nsim}{number of simulations to be performed. It is set at 100 by default.}

\item{clus}{the proportion of clusters to be used in parallel computing.}

\item{node}{number of nodes to be specifically tested for trends. It is NULL by default. Notice the node number must refer to the dichotomic version of the original tree, as produced by \code{RRphylo}.}

\item{cov}{the covariate values to be specified if the RR object has been created with covariate. As for \code{RRphylo}, \code{'cov'} must be as long as the number of nodes plus the number of tips of the tree, which can be obtained by running \code{RRphylo} on the covariate as well, and taking the vector of ancestral states and tip values to form the covariate (see the example below).}

\item{foldername}{the path of the folder where plots are to be found.}

\item{ConfInt}{if \code{TRUE}, the function returns 95\% confidence intervals for slopes of phenotype versus age, absolute rates versus age, and relative rates versus age regressions.}
}
\value{
The function returns a ‘list’ object including:

\strong{$rbt} for each branch of the tree, there are the age of the daughter node/tip (D.age), the age of the parental node (P.age), the \code{RRphylo} rates, and the distance from the tree root (age). If y is multivariate, it also includes the multiple rates for each y component.

\strong{$pbt} a data frame of phenotypic values and distance from the tree root for each node and tip.

\strong{$p.trend} results of phenotype versus age regression. It reports a p-value for the regression between the variables (p.real) a p-value computed contrasting the real slope to Brownian motion simulations (p.random), a global p-value corresponding to the least significant between p.real and p.random (p.value; see details for further explanations).

\strong{$rbt.rateA} results of absolute rate values versus age regression. It reports a p-value for the regression between the variables (p.real), a p-value computed contrasting the real slope to Brownian motion simulations (p.random), and a global p-value (p.value, corresponding to p.random; see details for further explanations).

\strong{$rbt.rateR} results of rate values versus age regression. It reports a p-value for the regression between the variables (p.real), a p-value computed contrasting the real slope to Brownian motion simulations (p.random), and a global p-value (p.value, corresponding to p.random; see details for further explanations).

\strong{$ConfInts} the 95\% confidence intervals around the slopes of phenotype versus age ($phenotype), absolute rates versus age ($abs.rate), and relative rates versus age ($rel.rate) regressions.

If the node argument is specified, the list also includes \strong{$p.trend.nodes}, \strong{$rbt.rateA.nodes}, \strong{$rbt.rateR.nodes}, which return the same results as above for each specified node. Finally, the \strong{$SMA} object contains the comparisons between slopes of regression lines of each pair of nodes, for all the regressions previously performed, under standard major axis regression.
}
\description{
This function searches for evolutionary trends in phenotypic mean and evolutionary rates .
}
\details{
The function simultaneously returns the regression of phenotypes and phenotypic evolutionary rates against age (in terms of distance from the tree root). As an output, plots for the absolute and relative rates regressions versus age are saved as a single .pdf file. In the plots, the 95\% confidence intervals of phenotypes and rates simulated under the Brownian motion for each node are plotted as shaded areas. The same is performed for the phenotype versus age regression. Regression slopes are reported for all regressions. To assess significance, slopes are compared to a family of simulated slopes (BMslopes, where the number of simulations is equal to \code{nsim}), generated as to show no phenotypic or rate trend (i.e. Brownian motion), using the \code{\link{setBM}} function. In the simulations, the phenotypic value at the root is the same as calculated for the original data by \code{RRphylo}. In addition, the Brownian rate is calculated for the original data and tree and set for all the Brownian motion simulations. Eventually, a p-value is obtained by comparing the real slope to BMslopes.
}
\examples{
 \donttest{
data("DataOrnithodirans")
DataOrnithodirans$treedino->treedino
DataOrnithodirans$massdino->massdino

# Extract Pterosaurs tree and data
  library(ape)
  extract.clade(treedino,748)->treeptero
  massdino[match(treeptero$tip.label,names(massdino))]->massptero
  massptero[match(treeptero$tip.label,names(massptero))]->massptero

# Case 1. "RRphylo" whitout accounting for the effect of a covariate
  RRphylo(tree=treeptero,y=log(massptero))->RRptero

 # Case 1.1. "search.trend" whitout indicating nodes to be tested for trends
   search.trend(RR=RRptero, y=log(massptero), nsim=100, clus=0.5,
   foldername=tempdir(),cov=NULL,ConfInt=FALSE,node=NULL)

 # Case 1.2. "search.trend" indicating nodes to be specifically tested for trends
    search.trend(RR=RRptero, y=log(massptero), nsim=100, node=143, clus=0.5,
    foldername=tempdir(),cov=NULL,ConfInt=FALSE)


# Case 2. "RRphylo" accounting for the effect of a covariate
 # "RRphylo" on the covariate in order to retrieve ancestral state values
   RRphylo(tree=treeptero,y=log(massptero))->RRptero
   c(RRptero$aces,log(massptero))->cov.values
   names(cov.values)<-c(rownames(RRptero$aces),names(massptero))
   RRphylo(tree=treeptero,y=log(massptero),cov=cov.values)->RRpteroCov

 # Case 2.1. "search.trend" whitout indicating nodes to be tested for trends
   search.trend(RR=RRpteroCov, y=log(massptero), nsim=100, clus=0.5,
   foldername=tempdir(),ConfInt=FALSE,cov=cov.values)

 # Case 2.2. "search.trend" indicating nodes to be specifically tested for trends
   search.trend(RR=RRpteroCov, y=log(massptero), nsim=100, node=143, clus=0.5,
   foldername=tempdir(),ConfInt=FALSE,cov=cov.values)}
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro Mondanaro, Marina Melchionna, Mirko Di Febbraro, Antonio Profico, Francesco Carotenuto
}
