% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.trend.R
\name{search.trend}
\alias{search.trend}
\title{Searching for evolutionary trends in phenotypes and rates}
\usage{
search.trend(RR,y,nsim=100,clus=.5,node=NULL,cov=NULL,foldername,ConfInt=FALSE)
}
\arguments{
\item{RR}{an object produced by \code{\link{RRphylo}}.}

\item{y}{the named vector (or matrix if multivariate) of phenotypes.}

\item{nsim}{number of simulations to be performed. It is set at 100 by default.}

\item{clus}{the proportion of clusters to be used in parallel computing.}

\item{node}{the node number of individual clades to be specifically tested and contrasted to each other. It is \code{NULL} by default. Notice the node number must refer to the dichotomic version of the original tree, as produced by \code{RRphylo}.}

\item{cov}{the covariate values to be specified if the RR object has been created using a  covariate for rates calculation.  As for \code{RRphylo}, \code{'cov'} must be as long as the number of nodes plus the number of tips of the tree, which can be obtained by running \code{RRphylo} on the covariate as well, and taking the vector of ancestral states and tip values to form the covariate (see the example below).}

\item{foldername}{the path of the folder where plots are to be found.}

\item{ConfInt}{if \code{TRUE}, the function returns 95\% confidence intervals around phenotypes and rates produced according to the Brownian motion model of evolution. It is \code{FALSE} by default.}
}
\value{
The function returns a ‘list’ object including:

\strong{$rbt} for each branch of the tree, there are the age of the daughter node/tip (D.age), the age of the parental node (P.age), the \code{RRphylo} rates, and the distance from the tree root (age). If y is multivariate, it also includes the multiple rates for each y vector.

\strong{$pbt} a data frame of phenotypic values and their distance from the tree root for each node (i.e. ancestral states) and tip of the tree.

\strong{$phenotypic.regression} results of phenotype versus age regression. It reports a p-value for the regression slope between the variables (p.real) a p-value computed contrasting the real slope to Brownian motion simulations (p.random). Only the latter should be inspected to assess significance.

\strong{$rate.regression} results of the rates (absolute values) versus age regression. It reports a p-value for the regression between the variables (p.real), a p-value computed contrasting the real slope to Brownian motion simulations (p.random). Only the latter should be inspected to assess significance.

\strong{$ConfInts} the 95\% confidence intervals around phenotypes and rates produced according to the Brownian motion model of evolution.

If specified, individual nodes are tested as the whole tree, the results are summarized in the objects \strong{$node.phenotypic.regression} and \strong{$node.rate.regression}. If more than one node is specified, the object  \strong{$group.comparison} reports the results obtained by comparing individual clades to each other.
}
\description{
This function searches for evolutionary trends in the phenotypic mean and the evolutionary rates for the entire tree and individual clades.
}
\details{
The function simultaneously returns the regression of phenotypes and phenotypic evolutionary rates against age (in terms of distance from the tree root) tested against Brownian motion simulations to assess signficance. It stores the rates (absolute values) versus age regression and the phenotype versus age regression plots as .pdf files. In the plots, the 95\% confidence intervals of phenotypes and rates simulated under the Brownian motion for each node are plotted as shaded areas. Regression lines are printed for all regressions. To assess significance, slopes are compared to a family of simulated slopes (BMslopes, where the number of simulations is equal to \code{nsim}), generated under the Brownian motion, using the \code{fastBM} function in the package \pkg{phytools}. Individual nodes are compared to the rest of the tree in different ways depending on whether phenotypes or rates versus age regressions are tested. With the former, the regression slopes for individual clades and the slope difference between clades is contrasted slopes obtained through Brownian motion simulations. For the latter, regression models are tested and contrasted to each other referring to estimated marginal means, by using the \code{emmeans} function in the package \pkg{emmeans}.
}
\examples{
 \donttest{
data("DataOrnithodirans")
DataOrnithodirans$treedino->treedino
DataOrnithodirans$massdino->massdino

# Extract Pterosaurs tree and data
  library(ape)
  extract.clade(treedino,748)->treeptero
  massdino[match(treeptero$tip.label,names(massdino))]->massptero
  massptero[match(treeptero$tip.label,names(massptero))]->massptero

# Case 1. "RRphylo" whitout accounting for the effect of a covariate
  RRphylo(tree=treeptero,y=log(massptero))->RRptero

 # Case 1.1. "search.trend" whitout indicating nodes to be tested for trends
   search.trend(RR=RRptero, y=log(massptero), nsim=100, clus=0.5,
   foldername=tempdir(),cov=NULL,ConfInt=FALSE,node=NULL)

 # Case 1.2. "search.trend" indicating nodes to be specifically tested for trends
    search.trend(RR=RRptero, y=log(massptero), nsim=100, node=143, clus=0.5,
    foldername=tempdir(),cov=NULL,ConfInt=FALSE)


# Case 2. "RRphylo" accounting for the effect of a covariate
 # "RRphylo" on the covariate in order to retrieve ancestral state values
   RRphylo(tree=treeptero,y=log(massptero))->RRptero
   c(RRptero$aces,log(massptero))->cov.values
   names(cov.values)<-c(rownames(RRptero$aces),names(massptero))
   RRphylo(tree=treeptero,y=log(massptero),cov=cov.values)->RRpteroCov

 # Case 2.1. "search.trend" whitout indicating nodes to be tested for trends
   search.trend(RR=RRpteroCov, y=log(massptero), nsim=100, clus=0.5,
   foldername=tempdir(),ConfInt=FALSE,cov=cov.values)

 # Case 2.2. "search.trend" indicating nodes to be specifically tested for trends
   search.trend(RR=RRpteroCov, y=log(massptero), nsim=100, node=143, clus=0.5,
   foldername=tempdir(),ConfInt=FALSE,cov=cov.values)}
}
\references{
Castiglione, S., Serio, C., Mondanaro, A., Di Febbraro, M., Profico, A., Girardi, G., & Raia, P. (2019) Simultaneous detection of macroevolutionary patterns in phenotypic means and rate of change with and within phylogenetic trees including extinct species. \emph{PLoS ONE}, 14: e0210101. https://doi.org/10.1371/journal.pone.0210101
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro Mondanaro, Marina Melchionna, Mirko Di Febbraro, Antonio Profico, Francesco Carotenuto
}
